inital refactor
