+++ .github/CONTRIBUTING.md
@@ -1,0 +1,4 @@
+## Contributing
+
+
+So if you're reading this, (and I'm not sure why) please feel free to contribute and try to point out problems in the code or things that could be clearer. Thanks!
