+++ .github/ISSUE_TEMPLATE.md
@@ -1,0 +1,11 @@
+## What's Breaking
+
+
+## My machine
+
+
+## How to reproduce
+
+
+## Suggested Fix
+
