+++ Makefile
@@ -4,15 +4,15 @@ CFLAGS = -Wall -g
 all: main
 
 main: clean
-	$(CC) $(CFLAGS) -o bookOrder main.c -lpthread
+	$(CC) $(CFLAGS) -o bookOrder bookOrder.c -lpthread
 
-run: clean main
+run: main
 	./bookOrder database.txt orders.txt categories.txt
 
 time: main
-	time ./bookOrder database.txt orders.txt categories.txt #> cpuTimeLog.txt
-	##cat cpuTimeLog.txt
+	time ./bookOrder database.txt orders.txt categories.txt
 
 clean:
+	clear
 	rm -f bookOrder
 	rm -f *.o
